// frontend/src/api/generated/personnel/personnel.zod.ts
export const PostApiPersonnelWithJsonBody = zod.object({
  "personalId": zod.string(),
  "fullName": zod.string(),
  "email": zod.string(),
  "allowedFactoryIds": zod.array(zod.number())
})
// ...
export const PutApiPersonnelIdWithJsonBody = zod.object({
  "personalId": zod.string(),
  "fullName": zod.string(),
  "email": zod.string(),
  "allowedFactoryIds": zod.array(zod.number())
})